+++ packages/widgets/audit-management-widget/src/lib/widget/api/types.ts
@@ -5,6 +5,7 @@ export type Sdk = ReturnType<typeof createWebSdk>;
 export type HttpClient = Sdk['httpClient'];
 
 export type Audit = {
+  ID?: string;
   id: string;
   userId: string;
   action: string;