+++ packages/widgets/audit-management-widget/test/mocks/rootMock.ts
@@ -14,8 +14,8 @@ export default `
     </descope-combo-box>
   </descope-container>
   <descope-grid bordered="true" column-reordering-allowed="true" data-id="audit-table" size="sm" st-host-height="300px" style="width:100%">
-  <descope-grid-text-column header="Occurred" path="occurredFormatted" resizable="true" sortable="false"></descope-grid-text-column>
-  <descope-grid-text-column header="User ID" path="userId" resizable="true" sortable="false"></descope-grid-text-column>
+    <descope-grid-text-column header="Occurred" path="occurredFormatted" resizable="true" sortable="false"></descope-grid-text-column>
+    <descope-grid-text-column header="User ID" path="userId" resizable="true" sortable="false"></descope-grid-text-column>
     <descope-grid-text-column header="Actor" path="actorId" resizable="true" sortable="false"></descope-grid-text-column>
     <descope-grid-text-column header="Login IDs" path="externalIds" resizable="true" sortable="false"></descope-grid-text-column>
     <descope-grid-text-column header="Remote Address" path="remoteAddress" resizable="true" sortable="false"></descope-grid-text-column>
@@ -26,6 +26,7 @@ export default `
       <descope-badge bordered="true" data-pattern="unknown" mode="default" size="xs" st-text-transform="capitalize"></descope-badge>
     </descope-grid-custom-column>
     <descope-grid-text-column header="Action" path="action" resizable="true" sortable="false"></descope-grid-text-column>
+    <descope-grid-item-details-column width="3.5em" flex-grow="0" frozen></descope-grid-item-details-column>
   </descope-grid>
 </descope-container>
 `;