+++ packages/sdks/core-js-sdk/CHANGELOG.md
@@ -2,6 +2,34 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.26.0](https://github.com/descope/descope-js/compare/core-js-sdk-2.25.1...core-js-sdk-2.26.0) (2024-09-29)
+
+
+### Features
+
+* Add passkeyOptions to all webauthn start calls ([#807](https://github.com/descope/descope-js/issues/807)) RELEASE ([a8e0909](https://github.com/descope/descope-js/commit/a8e09094f8afdb016f437a3bc3bb6c6586330840))
+
+## [2.25.1](https://github.com/descope/descope-js/compare/core-js-sdk-2.25.0...core-js-sdk-2.25.1) (2024-09-17)
+
+
+### Bug Fixes
+
+* Regression with encoding parameters ([#801](https://github.com/descope/descope-js/issues/801)) ([92105aa](https://github.com/descope/descope-js/commit/92105aaceb5fbfb1181acee90bd940680781d10b))
+
+## [2.25.0](https://github.com/descope/descope-js/compare/core-js-sdk-2.24.4...core-js-sdk-2.25.0) (2024-09-11)
+
+
+### Features
+
+* Get my tenants ([#794](https://github.com/descope/descope-js/issues/794)) ([e689111](https://github.com/descope/descope-js/commit/e689111153cb952b97a040d3e1c10ee90e0454bb))
+
+## [2.24.4](https://github.com/descope/descope-js/compare/core-js-sdk-2.24.3...core-js-sdk-2.24.4) (2024-09-03)
+
+
+### Bug Fixes
+
+* upgrade jwt decode to 4.0.0. ([#789](https://github.com/descope/descope-js/issues/789)) RELEASE ([19e2cfd](https://github.com/descope/descope-js/commit/19e2cfde2fd061110fec8918e211f89909553f8a))
+
 ## [2.24.3](https://github.com/descope/descope-js/compare/core-js-sdk-2.24.2...core-js-sdk-2.24.3) (2024-08-20)
 
 ## [2.24.2](https://github.com/descope/descope-js/compare/core-js-sdk-2.24.1...core-js-sdk-2.24.2) (2024-08-14)